refactor: separate screens
